Update 6
Adding Loading component that displays loading screen until the data request is complete.  Added data to the dataGrid display section to render the data as an unordered list of elements.  Added data into an object to reduce amount of props that need to be passed in the component call.  Added loading conditions to handle loading on/off.