Update 7
Added change between celsius and temperature button.  The button updates all temps in the ui on click.  Adjusted request so teh array is populated with temp, tempmin, temp max as well as the rest of the json data object to handle referincing when updating state between parent and child.  tempChange function is fed as props to the child datagrid component and used to handle the on click events.  The function updates the temperature attributes of the state and these are fed as props to the DataGrid component for rendering.  *Need to add conditionality to switch between F/C as now it is just switching to F then adding the calculation on every click.